requirements.txt
streamlit
supabase
cryptography
python-dotenv